project/requirements.txt
asgiref==3.5.0
crispy-bootstrap5==0.6
Django==3.2
django-crispy-forms==1.14.0
pytz==2022.1
sqlparse==0.4.2
typing-extensions==4.1.1
